requirements.txt
cryptography
pillow
pytest
pylint
black
